refactor: convert file browser from overlay drawer to inline panel
Replace the global FileBrowserDrawer overlay with an inline
FileBrowserPanel that renders as a flex sibling (like DiffPanel).
Each page manages its own width state and passes it down to ChatArea.
