+++ apps/frontend/src/pages/IssueDetailPage.tsx
@@ -9,10 +9,12 @@ import { CreateIssueDialog } from '@/components/kanban/CreateIssueDialog'
 import { MobileSidebar } from '@/components/kanban/MobileSidebar'
 import { useProject } from '@/hooks/use-kanban'
 import { useIsMobile } from '@/hooks/use-mobile'
+import { FILE_BROWSER_MIN_WIDTH } from '@/stores/file-browser-store'
 
 const SIDEBAR_WIDTH = 56
 const MIN_CHAT_WIDTH = 300
 const DEFAULT_DIFF_WIDTH = 360
+const DEFAULT_FILE_BROWSER_WIDTH = 360
 const DEFAULT_LIST_WIDTH = 232
 const MIN_LIST_WIDTH = 180
 const MAX_LIST_WIDTH = 400
@@ -28,6 +30,7 @@ export default function IssueDetailPage() {
   const { data: project, isLoading, isError } = useProject(projectId)
   const [showDiff, setShowDiff] = useState(false)
   const [diffWidth, setDiffWidth] = useState(DEFAULT_DIFF_WIDTH)
+  const [fileBrowserWidth, setFileBrowserWidth] = useState(DEFAULT_FILE_BROWSER_WIDTH)
   const [listWidth, setListWidth] = useState(DEFAULT_LIST_WIDTH)
   const isResizingList = useRef(false)
   const isMobile = useIsMobile()
@@ -80,6 +83,16 @@ export default function IssueDetailPage() {
       const listSpace = hideListPanel ? 0 : listWidth
       const maxWidth = viewport - SIDEBAR_WIDTH - listSpace - MIN_CHAT_WIDTH
       setDiffWidth(Math.min(Math.max(DIFF_MIN_WIDTH, w), maxWidth))
+    },
+    [hideListPanel, listWidth],
+  )
+
+  const handleFileBrowserWidthChange = useCallback(
+    (w: number) => {
+      const viewport = typeof window !== 'undefined' ? window.innerWidth : 1600
+      const listSpace = hideListPanel ? 0 : listWidth
+      const maxWidth = viewport - SIDEBAR_WIDTH - listSpace - MIN_CHAT_WIDTH
+      setFileBrowserWidth(Math.min(Math.max(FILE_BROWSER_MIN_WIDTH, w), maxWidth))
     },
     [hideListPanel, listWidth],
   )
@@ -140,6 +153,8 @@ export default function IssueDetailPage() {
               onToggleDiff={() => setShowDiff(v => !v)}
               onDiffWidthChange={handleDiffWidthChange}
               onCloseDiff={() => setShowDiff(false)}
+              fileBrowserWidth={fileBrowserWidth}
+              onFileBrowserWidthChange={handleFileBrowserWidthChange}
               showBackToList
             />
           ) :
